feat: refactor repository module
