+++ src/migrations.rs
@@ -1,13 +1,11 @@
 use std::sync::Arc;
 use libsql::Connection;
 use serde::{Deserialize, Serialize};
-use crate::version::VERSION;
 use semver::{Version, VersionReq};
 use tokio::sync::Mutex;
-use crate::faery::FaeryRepository;
-use crate::player;
-use crate::player::PlayerRepository;
-use crate::repository::{Repository, RepositoryError, RepositoryItem, RepositoryResult};
+use crate::repository::{RepositoryError, RepositoryItem, RepositoryResult};
+use crate::prelude::*;
+use crate::repository::player;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Migration {
