+++ src/prelude.rs
@@ -1,6 +1,8 @@
-pub use crate::faery::FaeryRepository;
+#![allow(dead_code, unused_imports, unused_variables, unused_mut, unused_unsafe)]
+pub use crate::repository::faery::FaeryRepository;
 pub use crate::repository::Repository;
 pub use crate::version::VERSION;
-pub use crate::email::EmailRepository;
-pub use crate::player::PlayerRepository;
-pub use crate::session::SessionRepository;
+pub use crate::repository::email::EmailRepository;
+pub use crate::repository::player::PlayerRepository;
+pub use crate::repository::player::PlayerData;
+pub use crate::repository::session::SessionRepository;
